% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateAll.R
\name{aggregateAll}
\alias{aggregateAll}
\title{Aggregate all hierarchy tables.}
\usage{
aggregateAll(.data, secHier, indHier, assHier, values)
}
\arguments{
\item{.data}{data.frame containing metrics to aggregate, along with Sector,
Industry, Asset, and Period columns.}

\item{secHier}{data.frame containing the sector hierarchy.}

\item{indHier}{data.frame containing the industry hierarchy.}

\item{assHier}{data.frame containing the asset hierarchy.}

\item{values}{character vector of column names defining the numeric variables to be aggregated.}
}
\value{
data.frame aggregated to all combinations across the hierarchies with
Sector_Level, Industry_Level, Asset_Level, and Group columns to denote the
aggregation level for each row.
}
\description{
Aggregate to all combinations of hierarchy tables.
}
\details{
Aggregates to all possible levels of the provided Sector, Industry, and Asset
hierarchy tables. Data must be provided at the lowest level of the hierarchies.

Values are aggregated by summing, with NA values disregarded. Otherwise a single
series' having NA for a Period will mean all higher level aggregates must also be
NA for that Period.

Hierarchy table columns must progress from least detailed to most detailed, i.e.
the left-most column represents the most aggregated level, and the right-most
column represents the most detailed level.
}
