% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregateHierarchy}
\alias{aggregateHierarchy}
\title{Aggregate up hierarchy.}
\usage{
aggregateHierarchy(.data, highLevel, lowLevel, values, keepLowLevel = FALSE)
}
\arguments{
\item{.data}{A data.frame}

\item{highLevel}{The high level column to which the data will be aggregated}

\item{lowLevel}{The low level column from which the data will be aggregated}

\item{values}{The columns that will be aggregated}

\item{keepLowLevel}{logical indicator}
}
\value{
A data.frame
}
\description{
Aggregate the data up an hierarchy from the given low level column to the given
high level column. Both columns must be categorical data.
}
\examples{
irisXL <- iris  \%>\% mutate(Extra=sample(c("one", "two"), nrow(iris), replace=TRUE))
res <- aggregateHierarchy(irisXL, "Extra", "Species",
   c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width"))
}
