% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregateLevels}
\alias{aggregateLevels}
\title{Aggregate levels of a categorical variable.}
\usage{
aggregateLevels(.data, column, values, levels, keepLowLevel = FALSE)
}
\arguments{
\item{.data}{A data.frame}

\item{column}{The column containing the categorical data}

\item{values}{The numeric column(s) that will be aggregated}

\item{levels}{A named character vector specifying the new value for any levels
that need updating (format new=old)}
}
\value{
A data.frame
}
\description{
Reduce the levels (unique values) of a categorical variable and aggregate
the reduced levels such that no duplicates occur.
The latter point is important because it is assumed the input data has no duplicates.
}
\examples{
aggregateLevels(iris, "Species",
     c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width"),
     c("alien"="setosa", "alien"="virginica"))
}
