% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregateWithHierarchyTable}
\alias{aggregateWithHierarchyTable}
\title{Aggregate up hierarchy table.}
\usage{
aggregateWithHierarchyTable(
  .data,
  column,
  hierarchyTable,
  targetLevel,
  values,
  withSubTotals = FALSE,
  keepOriginal = TRUE
)
}
\arguments{
\item{.data}{A data.frame}

\item{column}{The column name of the category for which the hierarchy is specified}

\item{hierarchyTable}{A data.frame outlining the hierarchy of the category with
columns ordered from lowest to highest detail}

\item{targetLevel}{A column in the hierarchy table}

\item{values}{The columns containing numeric values that will be aggregated}

\item{withSubTotals}{if TRUE subtotals will be added for each level between the
lowest level and the target level}

\item{keepOriginal}{if TRUE the aggregated rows will be added to \code{.data}, if
FALSE then the aggregated rows replace the original data}
}
\value{
A data.frame with same structure as \code{.data}, replaced or appended with aggregated rows
}
\description{
Aggregate the data up an hierarchy as specified in the given hierarchy table.
 The data is assumed to be at the lowest level within a category (e.g. sector)
 that resides in the given column. The numeric \code{values} are then
 aggregated to the given target level.
}
