% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregateWithHierarchyTableImp}
\alias{aggregateWithHierarchyTableImp}
\title{Aggregate up a hierarchy table (implementation).}
\usage{
aggregateWithHierarchyTableImp(
  .data,
  column,
  hierarchyTable,
  targetLevel,
  values,
  keepOriginal
)
}
\arguments{
\item{.data}{A data.frame}

\item{column}{The column name of the category for which the hierarchy is specified}

\item{hierarchyTable}{A data.frame outlining the hierarchy of the category}

\item{targetLevel}{A column in the hierarchy table}

\item{values}{The columns that will be aggregated}
}
\value{
A data.frame with the aggregated values added to it
}
\description{
Aggregate the data up an hierarchy as specified in the given hierarchy table.
 The data is assumed to be at the lowest level within a category (e.g. sector)
 that resides in the given column. The columns in the values argument are then
 aggregated to the given target level
}
