% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{calcAnnualCVMFlow}
\alias{calcAnnualCVMFlow}
\title{Calculate the annual CVM where the input time series represent flow values}
\usage{
calcAnnualCVMFlow(
  .data,
  lastCompleteYear,
  cypColumn = "CYP",
  pypColumn = "PYP"
)
}
\arguments{
\item{.data}{A data.frame with the Year, current year prices and previous year
prices}

\item{lastCompleteYear}{integer value for the last year in which all quarters were present}

\item{cypColumn}{Name of the column for the current year prices (defaults to CYP)}

\item{pypColumn}{Name of the column for the previous year prices (defaults to PYP)}
}
\value{
A data.frame where for each year there is a CVM value
}
\description{
Calculate the annual CVM where the input time series represent flow values
}
