% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{calcScalingFactor}
\alias{calcScalingFactor}
\title{Calculate the scaling factor (a function of the current year and previous year
prices). This is used to calculate the unconstrained CVM.}
\usage{
calcScalingFactor(
  .data,
  lastCompleteYear,
  cypColumn = "CYP",
  pypColumn = "PYP"
)
}
\arguments{
\item{.data}{A data.frame with the Year, current year prices and previous year
prices}

\item{lastCompleteYear}{integer value for the last year in which all quarters were present}

\item{cypColumn}{Name of the column for the current year prices (defaults to CYP)}

\item{pypColumn}{Name of the column for the previous year prices (defaults to PYP)}
}
\value{
A data.frame with the additional ScalingFactor column
}
\description{
Calculate the scaling factor (a function of the current year and previous year
prices). This is used to calculate the unconstrained CVM.
}
