% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{calcUnconstrainedCVM}
\alias{calcUnconstrainedCVM}
\title{Calculates the unconstrained CVM which is a function of the current year
prices, previous year prices and a scaling factor}
\usage{
calcUnconstrainedCVM(
  .data,
  lastCompleteYear,
  cypColumn = "CYP",
  pypColumn = "PYP"
)
}
\arguments{
\item{.data}{A data.frame with the Year, current year prices,  previous year
prices and scaling factors}

\item{lastCompleteYear}{integer value for the last year in which all quarters were present}

\item{cypColumn}{Name of the column for the current year prices (defaults to CYP)}

\item{pypColumn}{Name of the column for the previous year prices (defaults to PYP)}
}
\value{
A data.frame with the additional UnconstrainedCVM column
}
\description{
Calculates the unconstrained CVM which is a function of the current year
prices, previous year prices and a scaling factor
}
