% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{chain}
\alias{chain}
\title{Calculate constrained chain volume measure.}
\usage{
chain(
  .data,
  lastCompleteYear = 2016,
  cypColumn = "CYP",
  pypColumn = "PYP",
  chainType = c("Stock", "Flow"),
  benchType = 4
)
}
\arguments{
\item{.data}{A data.frame with the Year, current year prices and previous year
prices}

\item{lastCompleteYear}{integer value for the last year in which all quarters were present}

\item{cypColumn}{Name of the column for the current year prices (defaults to CYP)}

\item{pypColumn}{Name of the column for the previous year prices (defaults to PYP)}

\item{chainType}{Either Stock or Flow which determines the way the Annualised CVM
is calculated}
}
\value{
A data.frame as in the given input data with the addition of a CVM column
}
\description{
Calculate the constrained chained volume measure based on current year and
previous year prices (data is assumed to be quarterly).
}
\details{
CYP values must not contain NAs.

The series is benchmarked to the yearly figures. Benchmark cannot process
zero values so, where present, a small adjustment is made (+0.001) and removed
after benchmarking.
}
