% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chainAll.R
\name{chainAll}
\alias{chainAll}
\title{Chain all PYP/CYP pairs.}
\usage{
chainAll(
  .data,
  pairs = NULL,
  lastCompleteYear,
  parallelise = TRUE,
  benchType = 4
)
}
\arguments{
\item{.data}{A data.frame with one row per Sector/Asset/Industry with a "data" list-column
containing pairs of variables with previous and current year prices, and year information.}

\item{pairs}{A data.frame with PYP, CYP, and ChainType columns containing the variable name pairs for chaining.
If not provided it will created using any variable names ending with "PYP" or "CYP".}
}
\value{
The input data.frame with an addional "chained" list-column with new CVM variables.
}
\description{
Chain all PYP/CYP variables into CVM.
}
\details{
All input variables for should end in "CP" or "CVM". The CVM output variable
names are created using the CP variable name and appending with "CVM".
}
\section{Specification of pairs data.frame}{

The \code{pairs} data.frame associates the CYP and PYP variable names for each
output metric, e.g. associating GrossStockCYP with GrossStockPYP. The type
of chaining required (Stock or Flow) is also denoted. e.g.

\tabular{rrr}{
PYP \tab CYP \tab ChainType \cr
GrossStockPYP \tab GrossStockCYP \tab Stock \cr
NetStockPYP \tab NetStockCYP \tab Stock \cr
... \tab ... \tab ...
}
}

