% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrain.R
\name{constrain}
\alias{constrain}
\title{Constrain Series.}
\usage{
constrain(
  old = NULL,
  new,
  openPeriod = NULL,
  varName,
  pass = FALSE,
  zeroAdj = 0.001
)
}
\arguments{
\item{old}{data.frame with Period and a varName column containing the contraining values, usually from previously published data. Must contain at least one preceeding value from the openPeriod.}

\item{new}{data.frame with Period and a varName column containing the values to be contrained. Must contain the openPeriod and at least one preceeding value.}

\item{openPeriod}{String denoting the open period from which to extend old using new's growth rates}

\item{varName}{String denoting the stock/value to be constrained (must exist in both new and old)}

\item{zeroAdj}{An adjustment factor if zero values are present in new}
}
\value{
A data.frame with old up till the openPeriod and extended by the growth rates of new from the openPeriod onwards
}
\description{
Constrain one series to another.
}
\details{
The output series contains old values up until the openIndex. From the open index
onwards new values are calculated using the growth rates from new.

Series that contain zero values in the open period are adjusted to \code{zeroAdj}
before applying growth rates.

Data provided in old and new must each have contiguous Periods.

If pass = TRUE the new data.frame Period and varName columns are returned unmodified.
}
