% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrain.R
\name{constrainAll}
\alias{constrainAll}
\title{Constrain All Series.}
\usage{
constrainAll(
  old,
  new,
  openPeriod = NULL,
  varNames,
  pass = FALSE,
  zeroAdj = 0.001
)
}
\arguments{
\item{old}{data.frame of existing results, with one row per series and a nested \code{data} column containing Period and varNames columns. Series must be identified by Sector, Industry, and Asset columns.}

\item{new}{data.frame of new results, with one row per series and a nested \code{data} column containing Period and varNames columns. Series must be identified by Sector, Industry, and Asset columns.}

\item{openPeriod}{String denoting the open period from which to extend old using new's growth rates}

\item{varNames}{String vector denoting the columns to be constrained (must exist in both new and old's data list-column)}

\item{zeroAdj}{An adjustment factor if zero values are present in new}
}
\value{
The provided \code{old} data.frame with an additional "constrained" list-column containing the constrained values.
}
\description{
Constrain all series to previously published data
}
\details{
The output series contains old values up until the openIndex. From the open index
onwards new values are calculated using the growth rates from new.

Series that contain zero values in the open period are adjusted to \code{zeroAdj}
before applying growth rates.

old and new must contain Sector, Industry, and Asset columns to denote the series
and all series data must be provided as a nested list-column named "data".
}
