% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unchainOLD.R
\name{OLDunchain}
\alias{OLDunchain}
\title{Unchain Current Price Series.}
\usage{
OLDunchain(
  .data,
  cpColumn = "CP",
  cvmColumn = "CVM",
  deflatorColumn = NULL,
  deflatorMethod = c("implicit", "explicit")
)
}
\arguments{
\item{.data}{A data.frame with the current price, chained volume measures and
year information. The data.frame must also contain a Year column and optionally
a deflator column.}

\item{cpColumn}{The column name for the current price data ("CP" by default).}

\item{cvmColumn}{The column name for the CVM data ("CVM" by default). Can be ommitted if a deflator is provided.}

\item{deflatorColumn}{The column name for an optional deflator, see Details.}

\item{deflatorMethod}{String denoting deflator calculation method, "implicit"
or "explicit", see Details.}
}
\value{
The input data.frame with the addition of two columns (CYP and PYP).
}
\description{
Unchain current price series into current year prices and previous year prices.
}
\details{
CYP and PYP values are calculated from a deflator. If the deflatorMethod is
"implicit" a CVM column must be provided from which to calculate a deflator
(CP divided by CVM). If the deflatorMethod is "explicit" a deflator column
must be provided and its values are used directly.
}
