% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prorate.R
\name{prorateDownWithTimeSeries}
\alias{prorateDownWithTimeSeries}
\title{Prorate the specified aggregated category down in to the specified sub-categories
per the provided ratios}
\usage{
prorateDownWithTimeSeries(
  .data,
  column,
  aggregated_category,
  values,
  ratios,
  keepCategory = FALSE
)
}
\arguments{
\item{.data}{A data.frame}

\item{column}{The column containing the aggregated category}

\item{aggregated_category}{The name of the aggregated category}

\item{values}{Columns containing the values that need to be prorated}

\item{ratios}{A data.frame where each row is a time series of ratios for each
sub-category. Sub-categories are specified in a separate column. Ratio's must
sum to 1.}

\item{keepCategory}{boolean value for if we want to keep the aggregated category (TRUE)
or not (FALSE, default)}
}
\value{
A data.frame
}
\description{
Prorate the specified aggregated category down in to the specified sub-categories
per the provided ratios
}
