% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chainAll.R
\name{robustChain}
\alias{robustChain}
\title{Call chain with missing values present.}
\usage{
robustChain(
  .data,
  lastCompleteYear,
  cypColumn = "CYP",
  pypColumn = "PYP",
  chainType = c("Stock", "Flow"),
  benchType = 4
)
}
\arguments{
\item{.data}{A data.frame with the Year, current year prices and previous year
prices}

\item{lastCompleteYear}{integer value for the last year in which all quarters were present}

\item{cypColumn}{Name of the column for the current year prices (defaults to CYP)}

\item{pypColumn}{Name of the column for the previous year prices (defaults to PYP)}

\item{chainType}{Either Stock or Flow which determines the way the Annualised CVM
is calculated}
}
\value{
A data.frame as in the given input data with the addition of a CVM column
}
\description{
Call the chain function whilst dealing with leading missing values, and all zero values.
}
\details{
For metrics dealing with \emph{change} in quantities, e.g. "TotalChangesInVolume"
the first year in the series is often NA. chain is called without the inital
year of data and NAs are added afterwards.

Where all the values in the given cypColumn are zero, the series is returned
without calling chain.
}
