% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unchain.R
\name{unchain}
\alias{unchain}
\title{Unchain Current Price Series.}
\usage{
unchain(.data1, cpColumn = "CP", flow_stock = NULL, refYear = refYear)
}
\arguments{
\item{.data1}{A data.frame with the current price, chained volume measures and
year information. The data.frame must also contain a Year column}

\item{cpColumn}{The column name for the current price data ("CP" by default).}

\item{cvmColumn}{The column name for the CVM data ("CVM" by default). Can be ommitted if a deflator is provided.}
}
\value{
The input data.frame with the addition of two columns (CYP and PYP).
}
\description{
https://www.ons.gov.uk/economy/nationalaccounts/uksectoraccounts/methodologies/chainlinkingmethodsusedwithintheuknationalaccounts
}
\details{
This function creates CYPs and PYPs for pairs of CP and 'CVM' series
}
