% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unchainAll.R
\name{unchainAll}
\alias{unchainAll}
\title{Unchain all CP/CVM pairs.}
\usage{
unchainAll(.data, pairs, refYear, parallelise = TRUE)
}
\arguments{
\item{.data}{A data.frame with one row per Sector/Asset/Industry with a "data" list-column
containing current price, chained volume measures, Year columns, and optionally a price index column.}

\item{pairs}{A data.frame with CP and CVM columns containing the variable name pairs for unchaining.
If deflatorMethod is "explicit" the CVM column is not used.}

\item{parallelise}{If TRUE (the default) processing will run on a local parallel
cluster using all CPU cores except one.}

\item{deflatorColumn}{The column name for an optional deflator when deflatorMethod is "explicit".}

\item{deflatorMethod}{String denoting deflator calculation method, "implicit"
or "explicit", see Details.}
}
\value{
The input data.frame with an addional "unchained" list-column with CP
varnames appended with "CYP", and "PYP".
}
\description{
Unchain all CP/CVM variables into Current Year Prices and Previous Year Prices
using parallel processing.
}
\details{
The default deflatorMethod of "implicit" will infer a deflator from the CP and CVM
series (CP divided by CVM). If the "explicit" method is used, a deflatorColumn
must be provided and the CVM series are not used. See \code{unchain} for details.
}
